+++ data/macros/open-spell-importer.js
@@ -1,0 +1,6 @@
+/**
+ ****************************************************************
+ * This macro can be used to open the spell importer
+ ***************************************************************/
+
+new shadowdark.apps.SpellImporterSD().render(true);