+++ system/src/apps/_module.mjs
@@ -1,7 +1,10 @@
 // Property selector applications
+export {default as CharacterGeneratorSD} from "./CharacterGeneratorSD.mjs";
 export {default as EffectPanelSD} from "./EffectPanelSD.mjs";
 export {default as GemBagSD} from "./GemBagSD.mjs";
+export {default as ItemImporterSD} from "./ItemImporterSD.mjs";
 export {default as LightSourceTrackerSD} from "./LightSourceTrackerSD.mjs";
 export {default as MonsterImporterSD} from "./MonsterImporterSD.mjs";
 export {default as NpcAttackRangesSD} from "./NpcAttackRangesSD.mjs";
 export {default as ShadowdarklingImporterSD} from "./ShadowdarklingImporterSD.mjs";
+export {default as SpellImporterSD} from "./SpellImporterSD.mjs";