+++ system/src/socket.mjs
@@ -1,0 +1,54 @@
+export default function listenOnSocket() {
+
+	game.socket.on("system.shadowdark", event => {
+		if (event.type === "createCharacter") {
+			// only the GM should handle this event
+			if (!game.user.isGM) return;
+
+			shadowdark.apps.CharacterGeneratorSD.createActorFromData(
+				event.payload.characterData,
+				event.payload.characterItems,
+				event.payload.userId
+			);
+		}
+
+		if (event.type === "dropLightSourceOnScene" && game.user.isGM) {
+			game.shadowdark.lightSourceTracker.dropLightSourceOnScene(
+				event.data.item,
+				event.data.itemOwner,
+				event.data.actorData,
+				event.data.dropData,
+				event.data.speaker
+			);
+		}
+
+		if (event.type === "openNewCharacter") {
+			if (event.payload.userId === game.userId) {
+				const actor = game.actors.get(event.payload.actorId);
+
+				if (actor) actor.sheet.render(true);
+
+				return ui.notifications.info(
+					game.i18n.localize("SHADOWDARK.apps.character-generator.success"),
+					{permanent: false}
+				);
+			}
+		}
+
+		if (event.type === "pickupLightSourceFromScene" && game.user.isGM) {
+			game.shadowdark.lightSourceTracker.pickupLightSourceFromScene(
+				event.data.character,
+				event.data.lightActor,
+				event.data.lightToken,
+				event.data.speaker
+			);
+		}
+
+		if (event.type === "toggleLightSource" && game.user.isGM) {
+			game.shadowdark.lightSourceTracker.toggleLightSource(
+				event.data.actor,
+				event.data.item
+			);
+		}
+	});
+}